added try catch to delete expressions

+++ src/actions/Form/Input/handleInputSubmit.ts
@@ -60,7 +60,9 @@ const handleInputSubmit = async (interaction: ButtonInteraction | any, questionI
 	await interaction.deferReply({ ephemeral: true });
 
 	if (!_formEntry || !_formEntry.form) {
-		await message.delete();
+		if (message.deletable) {
+			await message.delete();
+		}
 		return await interaction.followUp({
 			content: ` You can't use the form. Looks like the form is deleted`,
 			ephemeral: true
