+++ src/actions/Select/handleSelectInput.ts
@@ -12,7 +12,9 @@ const handleSelectInput = async (interaction: SelectMenuInteraction, questionId:
 	await interaction.deferReply({ ephemeral: true });
 
 	if (!_formEntry || !_formEntry.form) {
-		await message.delete();
+		if (message.deletable) {
+			await message.delete();
+		}
 		return await interaction.followUp({
 			content: ` You can't use the form. Looks like the form is deleted`,
 			ephemeral: true
